+++ cli/src/commit_templater.rs
@@ -138,6 +138,7 @@ pub struct CommitTemplateLanguage<'repo> {
     repo: &'repo dyn Repo,
     path_converter: &'repo RepoPathUiConverter,
     workspace_name: WorkspaceNameBuf,
+    env_vars: &'repo HashMap<String, String>,
     // RevsetParseContext doesn't borrow a repo, but we'll need 'repo lifetime
     // anyway to capture it to evaluate dynamically-constructed user expression
     // such as `revset("ancestors(" ++ commit_id ++ ")")`.
@@ -162,6 +163,7 @@ impl<'repo> CommitTemplateLanguage<'repo> {
         repo: &'repo dyn Repo,
         path_converter: &'repo RepoPathUiConverter,
         workspace_name: &WorkspaceName,
+        env_vars: &'repo HashMap<String, String>,
         revset_parse_context: RevsetParseContext<'repo>,
         id_prefix_context: &'repo IdPrefixContext,
         immutable_expression: Arc<UserRevsetExpression>,
@@ -182,6 +184,7 @@ impl<'repo> CommitTemplateLanguage<'repo> {
             repo,
             path_converter,
             workspace_name: workspace_name.to_owned(),
+            env_vars,
             revset_parse_context,
             id_prefix_context,
             immutable_expression,
@@ -202,6 +205,10 @@ impl<'repo> CommitTemplateLanguage<'repo> {
 
 impl<'repo> TemplateLanguage<'repo> for CommitTemplateLanguage<'repo> {
     type Property = CommitTemplatePropertyKind<'repo>;
+
+    fn env_vars(&self) -> &HashMap<String, String> {
+        self.env_vars
+    }
 
     fn settings(&self) -> &UserSettings {
         self.repo.base_repo().settings()
@@ -3079,6 +3086,7 @@ mod tests {
         revset_aliases_map: RevsetAliasesMap,
         template_aliases_map: TemplateAliasesMap,
         immutable_expression: Arc<UserRevsetExpression>,
+        env_vars: HashMap<String, String>,
         extra_functions: HashMap<&'static str, BuildFunctionFn>,
     }
 
@@ -3103,6 +3111,7 @@ mod tests {
                 revset_aliases_map: RevsetAliasesMap::new(),
                 template_aliases_map: TemplateAliasesMap::new(),
                 immutable_expression: RevsetExpression::none(),
+                env_vars: HashMap::new(),
                 extra_functions: HashMap::new(),
             }
         }
@@ -3137,6 +3146,7 @@ mod tests {
                 self.test_workspace.repo.as_ref(),
                 &self.path_converter,
                 self.test_workspace.workspace.workspace_name(),
+                &self.env_vars,
                 revset_parse_context,
                 &self.id_prefix_context,
                 self.immutable_expression.clone(),