+++ cli/src/operation_templater.rs
@@ -67,6 +67,7 @@ pub trait OperationTemplateEnvironment {
 pub struct OperationTemplateLanguage {
     repo_loader: RepoLoader,
     current_op_id: Option<OperationId>,
+    env_vars: HashMap<String, String>,
     build_fn_table: OperationTemplateLanguageBuildFnTable,
     cache_extensions: ExtensionsMap,
 }
@@ -77,6 +78,7 @@ impl OperationTemplateLanguage {
     pub fn new(
         repo_loader: &RepoLoader,
         current_op_id: Option<&OperationId>,
+        env_vars: &HashMap<String, String>,
         extensions: &[impl AsRef<dyn OperationTemplateLanguageExtension>],
     ) -> Self {
         let mut build_fn_table = OperationTemplateLanguageBuildFnTable::builtin();
@@ -93,6 +95,7 @@ impl OperationTemplateLanguage {
             // Clone these to keep lifetime simple
             repo_loader: repo_loader.clone(),
             current_op_id: current_op_id.cloned(),
+            env_vars: env_vars.clone(),
             build_fn_table,
             cache_extensions,
         }
@@ -101,6 +104,10 @@ impl OperationTemplateLanguage {
 
 impl TemplateLanguage<'static> for OperationTemplateLanguage {
     type Property = OperationTemplateLanguagePropertyKind;
+
+    fn env_vars(&self) -> &HashMap<String, String> {
+        &self.env_vars
+    }
 
     fn settings(&self) -> &UserSettings {
         self.repo_loader.settings()